+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/blocks/EssenceBlockRed.java
@@ -4,6 +4,7 @@ import com.mojang.serialization.MapCodec;
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
 import com.telepathicgrunt.the_bumblezone.blocks.blockentities.EssenceBlockEntity;
 import com.telepathicgrunt.the_bumblezone.bossbars.ServerEssenceEvent;
+import com.telepathicgrunt.the_bumblezone.configs.BzGeneralConfigs;
 import com.telepathicgrunt.the_bumblezone.items.essence.EssenceOfTheBees;
 import com.telepathicgrunt.the_bumblezone.modinit.BzSounds;
 import com.telepathicgrunt.the_bumblezone.modinit.BzStats;
@@ -73,7 +74,7 @@ public class EssenceBlockRed extends EssenceBlock {
 
     @Override
     public int getEventTimeFrame() {
-        return 11200;
+        return BzGeneralConfigs.redArenaTimeFrameInTicks;
     }
 
     @Override